+++ lib/heroku/forward/backends/unicorn.rb
@@ -6,7 +6,7 @@ module Heroku
         
         def initialize(options = {})
           @application = options[:application]
-          @socket = options[:socket] || tmp_filename('unicorn', '.sock')
+          @socket = options[:socket] || Heroku::Forward::Utils::Dir.tmp_filename('unicorn', '.sock')
           @env = options[:env] || 'development'
           @config_file = options[:config_file]
         end
@@ -37,11 +37,6 @@ module Heroku
         end
         
         private
-        
-        # Borrowed from ruby 1.9's Dir::Tmpname.make_tmpname for 1.8.7-compatibility.
-        def tmp_filename(prefix, suffix)
-          File.join Dir.tmpdir, "#{prefix}#{Time.now.strftime("%Y%m%d")}-#{$$}-#{rand(0x100000000).to_s(36)}#{suffix}"
-        end
         
         def check!
           raise Heroku::Forward::Errors::MissingBackendOptionError.new('application') unless @application && @application.length > 0